refactor: rename to Accordion
