+++ docs/content/2.elements/7.accordion.md
@@ -1,16 +1,16 @@
 ---
 github: true
-description: Display toggleable accordion panels for interactive content display and hiding.
+description: Display toggleable accordionpanels for interactive content display and hiding.
 headlessui:
-  label: 'Disclosure'
-  to: 'https://headlessui.com/vue/disclosure'
+  label: 'Accordion '
+  to: 'https://headlessui.com/vue/accordion'
 navigation:
   badge: 'New'
 ---
 
 ## Usage
 
-Pass an array of arrays to the `items` prop of the Disclosure component. Each item can have the following properties:
+Pass an array of arrays to the `items` prop of the Accordion  component. Each item can have the following properties:
 
 - `label` - The label of the item.
 - `leadingIcon` - An optional icon displayed before the label.
@@ -21,7 +21,7 @@ Pass an array of arrays to the `items` prop of the Disclosure component. Each it
 
 ::component-example
 #default
-:disclosure-example-basic
+:accordion-example-basic
 
 #code
 ```vue
@@ -43,14 +43,14 @@ const items = ref([
 </script>
 
 <template>
-  <UDisclosure :items="items" />
+  <UAccordion :items="items" />
 </template>
 ```
 ::
 
 ### Color 
 
-Use the `color` prop to change the color of the disclosure. The default color is `primary`.
+Use the `color` prop to change the color of the accordion. The default color is `primary`.
 
 ::component-card
 ---
@@ -71,7 +71,7 @@ props:
 
 ### Size
 
-Use the `size` prop to change the size of the disclosure. The default size is `md`.
+Use the `size` prop to change the size of the accordion. The default size is `md`.
 
 ::component-card
 ---
@@ -92,7 +92,7 @@ props:
 
 ### Icons open / close
 
-Use the `openIcon` and `closeIcon` props to change the icons used to open and close or change them globally in `ui.disclosure.default.openIcon` and `ui.disclosure.default.closeIcon`.
+Use the `openIcon` and `closeIcon` props to change the icons used to open and close or change them globally in `ui.accordion.default.openIcon` and `ui.accordion.default.closeIcon`.
 
 You can also set it to `null` to hide the icons.
 
@@ -111,6 +111,9 @@ baseProps:
 props:
   openIcon: 'i-heroicons-plus'
   closeIcon: 'i-heroicons-minus'
+excludedProps:
+  - openIcon
+  - closeIcon
 ---
 ::
 
@@ -136,7 +139,7 @@ baseProps:
       content: "Lorem ipsum dolor sit amet, consectetur adipiscing elit"
 ---
 
-Disclosure
+Accordion 
 ::
 
 #code
@@ -153,7 +156,7 @@ const items = ref([
 </script>
 
 <template>
-  <UDisclosure :items="items" />
+  <UAccordion :items="items" />
 </template>
 ```
 ::
@@ -176,6 +179,8 @@ baseProps:
       content: "Lorem ipsum dolor sit amet, consectetur adipiscing elit"
 props:
   defaultOpenAll: true
+excludedProps:
+  - defaultOpenAll
 ---
 ::
 
@@ -185,11 +190,11 @@ You can use slots to customize the button and content items of the table.
 
 ### `button`
 
-Use the `#button` to customize the button of the disclosure. The slot contains **leading icon**, **index of item** ,**title** and **open / close icon**.
+Use the `#button` to customize the button of the accordion. The slot contains **leading icon**, **index of item** ,**title** and **open / close icon**.
 
 ::component-example
 #default
-:disclosure-example-button-slot
+:accordion-example-button-slot
 
 #code
 ```vue
@@ -207,7 +212,7 @@ const items = ref([
 </script>
 
 <template>
-  <UDisclosure :items="items">
+  <UAccordion :items="items">
     <template #button="{ title ,index ,leadingIcon, isOpen }">
       <div class="flex justify-start items-center border-b w-full pt-1 pb-3">
         <span class="bg-primary-200 rounded-full text-primary-800 p-[8px] mr-4 flex justify-center">
@@ -220,17 +225,12 @@ const items = ref([
         <div class="flex justify-between items-center w-full">
           <div>{{ index+1 }}.  {{ title }}</div>
           <div>
-            <UIcon
-              :name="{
-                'i-heroicons-plus': !isOpen,
-                'i-heroicons-minus': isOpen,
-              }"
-            />
+            <UIcon :name="isOpen ? 'i-heroicons-minus' : 'i-heroicons-plus'" />
           </div>
         </div>
       </div>
     </template>
-  </UDisclosure>
+  </UAccordion>
 </template>
 ```
 ::
@@ -241,7 +241,7 @@ Use the `#item-<index>-content` to customize a specific item content.
 
 ::component-example
 #default
-:disclosure-example-content-slot
+:accordion-example-content-slot
 
 #code
 ```vue
@@ -270,7 +270,7 @@ const subItems = ref([...]);
 </script>
 
 <template>
-  <UDisclosure :items="items" all-opend>
+  <UAccordion :items="items" all-opend>
     <template #item-1-content>
       <div class="flex justify-start items-center w-full pt-1 pb-3">
         <div class="flex flex-col justify-center items-start ml-2">
@@ -283,7 +283,7 @@ const subItems = ref([...]);
         </div>
       </div>
 
-      <UDisclosure :items="subItems" color="emerald" />
+      <UAccordion :items="subItems" color="emerald" />
     </template>
 
     <template #item-2-content="{ content }">
@@ -299,7 +299,7 @@ const subItems = ref([...]);
         <code>$ pnpm i -D @nuxthq/ui</code>
       </div>
     </template>
-  </UDisclosure>
+  </UAccordion>
 </template>
 ```
 ::
